# Backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Management API",
    "description": "Backend API services for Project Management System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/manage/{entity_type}": {
      "post": {
        "tags": [
          "Projects & Tasks Management"
        ],
        "summary": "Unified Create API (Path variable differentiator: 'project' or 'task')",
        "description": "Creates a new Project or Task depending on the '{entity_type}' path variable ('project' or 'task'). Enforces PBAC.",
        "operationId": "manage_create_entity_api_manage__entity_type__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Type"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Manage Create Entity Api Manage  Entity Type  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/manage/{entity_type}/{entity_id}": {
      "post": {
        "tags": [
          "Projects & Tasks Management"
        ],
        "summary": "Unified Update API (Path variable differentiator: 'project' or 'task' + entity_id)",
        "description": "Updates an existing Project or Task depending on '{entity_type}' path variable and '{entity_id}'. Enforces PBAC.",
        "operationId": "manage_update_entity_api_manage__entity_type___entity_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Type"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Manage Update Entity Api Manage  Entity Type   Entity Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects & Tasks Management"
        ],
        "summary": "Unified Delete API (Path variable differentiator: 'project' or 'task' + entity_id)",
        "description": "Deletes an existing Project or Task. Enforces PBAC.",
        "operationId": "manage_delete_entity_api_manage__entity_type___entity_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Type"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Manage Delete Entity Api Manage  Entity Type   Entity Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/lov": {
      "get": {
        "tags": [
          "Projects & Tasks Management"
        ],
        "summary": "Get Project LOV Options (Authenticated)",
        "description": "Fetch LOV options for project creation and dropdowns.",
        "operationId": "get_project_lovs_projects_lov_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectLOVResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/projects": {
      "get": {
        "tags": [
          "Projects & Tasks Management"
        ],
        "summary": "List user projects",
        "description": "Fetch all projects where current user is owner or member.",
        "operationId": "list_projects_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "type": "array",
                  "title": "Response List Projects Projects Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Projects & Tasks Management"
        ],
        "summary": "Create a new project",
        "description": "Direct route to create a new project.",
        "operationId": "create_project_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/projects/{project_id}": {
      "post": {
        "tags": [
          "Projects & Tasks Management"
        ],
        "summary": "Update a project by ID",
        "description": "Direct route to update an existing project (Enforces PBAC).",
        "operationId": "update_project_projects__project_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects & Tasks Management"
        ],
        "summary": "Delete a project by ID",
        "description": "Direct route to delete an existing project (Enforces PBAC).",
        "operationId": "delete_project_projects__project_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Projects & Tasks Management"
        ],
        "summary": "Get project details by ID",
        "description": "Fetch a single project by ID (Enforces PBAC Project Visibility).",
        "operationId": "get_project_by_id_projects__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "post": {
        "tags": [
          "Projects & Tasks Management"
        ],
        "summary": "Create a new task",
        "description": "Direct route to create a new task (Enforces PBAC Task Creation).",
        "operationId": "create_task_tasks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Projects & Tasks Management"
        ],
        "summary": "List all tasks",
        "description": "Fetch tasks strictly enforcing PBAC policy rules at the database query level.",
        "operationId": "list_tasks_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "assignee_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assignee Id"
            }
          },
          {
            "name": "status_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Id"
            }
          },
          {
            "name": "priority_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response List Tasks Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProjectResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Project created successfully"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectResponse"
          }
        },
        "type": "object",
        "required": [
          "project"
        ],
        "title": "CreateProjectResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreate": {
        "properties": {
          "project_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Project Name",
            "description": "Name of the project (Mandatory)"
          },
          "status_id": {
            "type": "string",
            "format": "uuid",
            "title": "Status Id",
            "description": "Status ID (Mandatory)"
          },
          "priority_id": {
            "type": "string",
            "format": "uuid",
            "title": "Priority Id",
            "description": "Priority ID (Mandatory)"
          },
          "project_type_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Type Id",
            "description": "Project Type ID (Mandatory)"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id",
            "description": "Category ID (Mandatory)"
          },
          "planned_start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Planned Start Date",
            "description": "Planned start date (Mandatory)"
          },
          "planned_end_date": {
            "type": "string",
            "format": "date-time",
            "title": "Planned End Date",
            "description": "Planned end date (Mandatory)"
          },
          "estimated_duration": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Estimated Duration",
            "description": "Estimated duration in days (Mandatory)"
          },
          "project_description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Description",
            "description": "Description of the project (Optional)"
          },
          "actual_start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Start Date",
            "description": "Actual start date (Optional)"
          },
          "actual_end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual End Date",
            "description": "Actual end date (Optional)"
          }
        },
        "type": "object",
        "required": [
          "project_name",
          "status_id",
          "priority_id",
          "project_type_id",
          "category_id",
          "planned_start_date",
          "planned_end_date",
          "estimated_duration"
        ],
        "title": "ProjectCreate",
        "description": "Schema for creating a project."
      },
      "ProjectLOVResponse": {
        "properties": {
          "statuses": {
            "items": {
              "$ref": "#/components/schemas/LOVItem"
            },
            "type": "array",
            "title": "Statuses"
          },
          "priorities": {
            "items": {
              "$ref": "#/components/schemas/LOVItem"
            },
            "type": "array",
            "title": "Priorities"
          },
          "project_types": {
            "items": {
              "$ref": "#/components/schemas/LOVItem"
            },
            "type": "array",
            "title": "Project Types"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/LOVItem"
            },
            "type": "array",
            "title": "Categories"
          },
          "task_types": {
            "items": {
              "$ref": "#/components/schemas/LOVItem"
            },
            "type": "array",
            "title": "Task Types",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "statuses",
          "priorities",
          "project_types",
          "categories"
        ],
        "title": "ProjectLOVResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}